Move gesture-area controls to a dedicated screen
The four Gesture Areas sliders no longer expand on the main dashboard; the main
screen now shows a compact entry card (parallel to the App compatibility entry)
that opens a dedicated Gesture areas screen.

Main screen:
- SectionHeader("Gesture areas") is kept; under it a GestureAreasEntryCard with
  the same Surface/extraLarge/surfaceContainerHigh visual language as
  CompatEntryCard — full-width click target, bodyMedium description, trailing ›.

Dedicated GestureAreasScreen (new ui/GestureAreasScreen.kt):
- Mirrors CompatibilityScreen: Scaffold + TopAppBar (back arrow, "Gesture areas"
  title, surface-colored), scrollable column, short info card, then the detailed
  settings card with the existing Back / Home & Recents grouping, labels, hints,
  and immediate-persistence slider behavior unchanged.
- GestureAreasCard, AreaSectionHeader, PercentSlider, SensitivitySlider are moved
  here (private) from MainActivity.kt, so MainActivity returns to being the
  dashboard + lightweight navigation.

Navigation: a mutually-exclusive AppScreen enum (MAIN / GESTURE_AREAS /
COMPATIBILITY) drives the existing local Compose navigation — no Navigation
Compose dependency. BackHandler returns to MAIN; system Back does the same;
CompatibilityScreen's internal App Picker navigation is unchanged.

No runtime gesture geometry, settings, DataStore keys, defaults, clamping,
computeGestureZoneLayout, EdgeOverlayController, indicator sizing, or overlay
architecture is modified — UI/navigation restructuring only. New strings:
gesture_areas_screen_title, gesture_areas_entry_desc.
