+++ app/src/main/java/com/ogesture/MainActivity.kt
@@ -44,13 +44,11 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.Slider
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -59,7 +57,6 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.graphics.Path
 import androidx.compose.ui.graphics.StrokeJoin
 import androidx.compose.ui.graphics.drawscope.Stroke
@@ -76,7 +73,6 @@ import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.withLink
 import androidx.compose.ui.unit.dp
-import com.ogesture.data.GestureZoneSettings
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LocalLifecycleOwner
@@ -86,11 +82,20 @@ import com.ogesture.ui.AccessibilityConsentDialog
 import com.ogesture.ui.AccessibilityStatus
 import com.ogesture.ui.CompatEntryCard
 import com.ogesture.ui.CompatibilityScreen
+import com.ogesture.ui.GestureAreasEntryCard
+import com.ogesture.ui.GestureAreasScreen
 import com.ogesture.ui.MainViewModel
 import com.ogesture.ui.PRIVACY_POLICY_URL
 import com.ogesture.ui.SetupCard
 import com.ogesture.ui.theme.OgestureTheme
 import kotlinx.coroutines.delay
+
+/**
+ * Lightweight, mutually-exclusive navigation destinations for the app's local Compose
+ * navigation (no Navigation Compose dependency). At most one secondary screen is active at
+ * a time; [AppScreen.MAIN] is the dashboard, the other two are reached via entry cards.
+ */
+enum class AppScreen { MAIN, GESTURE_AREAS, COMPATIBILITY }
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -101,12 +106,24 @@ class MainActivity : ComponentActivity() {
         window.isNavigationBarContrastEnforced = false
         setContent {
             OgestureTheme {
-                var showCompat by rememberSaveable { mutableStateOf(false) }
-                if (showCompat) {
-                    BackHandler { showCompat = false }
-                    CompatibilityScreen(onBack = { showCompat = false })
-                } else {
-                    MainScreen(onOpenCompat = { showCompat = true })
+                // Lightweight local navigation — no Navigation Compose dependency. A single
+                // mutually-exclusive destination so two secondary screens can't be active at once.
+                var screen by rememberSaveable { mutableStateOf(AppScreen.MAIN) }
+                when (screen) {
+                    AppScreen.GESTURE_AREAS -> {
+                        BackHandler { screen = AppScreen.MAIN }
+                        GestureAreasScreen(onBack = { screen = AppScreen.MAIN })
+                    }
+                    AppScreen.COMPATIBILITY -> {
+                        BackHandler { screen = AppScreen.MAIN }
+                        CompatibilityScreen(onBack = { screen = AppScreen.MAIN })
+                    }
+                    AppScreen.MAIN -> {
+                        MainScreen(
+                            onOpenGestureAreas = { screen = AppScreen.GESTURE_AREAS },
+                            onOpenCompat = { screen = AppScreen.COMPATIBILITY },
+                        )
+                    }
                 }
             }
         }
@@ -115,7 +132,11 @@ class MainActivity : ComponentActivity() {
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-private fun MainScreen(onOpenCompat: () -> Unit, viewModel: MainViewModel = viewModel()) {
+private fun MainScreen(
+    onOpenGestureAreas: () -> Unit,
+    onOpenCompat: () -> Unit,
+    viewModel: MainViewModel = viewModel(),
+) {
     val context = LocalContext.current
     val masterEnabled by viewModel.masterEnabled.collectAsState()
 
@@ -225,7 +246,7 @@ private fun MainScreen(onOpenCompat: () -> Unit, viewModel: MainViewModel = view
             GesturesCard()
 
             SectionHeader(stringResource(R.string.gesture_areas_title))
-            GestureAreasCard(viewModel = viewModel)
+            GestureAreasEntryCard(onClick = onOpenGestureAreas)
 
             SectionHeader(stringResource(R.string.compat_title))
             CompatEntryCard(onClick = onOpenCompat)
@@ -372,170 +393,6 @@ private fun GesturesCard() {
                 )
             }
         }
-    }
-}
-
-/**
- * Separate card for the four user-tunable gesture-area settings. Sliders use local Compose
- * state while dragging (snapped to the allowed step) and persist on onValueChangeFinished, so
- * dragging a thumb doesn't spam DataStore writes or trigger dozens of overlay rebuilds per
- * second; the committed value updates the active gesture windows immediately.
- */
-@Composable
-private fun GestureAreasCard(viewModel: MainViewModel) {
-    val settings by viewModel.gestureZoneSettings.collectAsState()
-
-    Card(
-        shape = MaterialTheme.shapes.extraLarge,
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surfaceContainerHigh,
-        ),
-        elevation = CardDefaults.cardElevation(defaultElevation = 0.dp),
-    ) {
-        Column(modifier = Modifier.padding(vertical = 8.dp)) {
-            Text(
-                text = stringResource(R.string.gesture_areas_subtitle),
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                modifier = Modifier.padding(horizontal = 16.dp, vertical = 4.dp),
-            )
-
-            AreaSectionHeader(stringResource(R.string.areas_back_header))
-            PercentSlider(
-                label = stringResource(R.string.areas_back_height_label),
-                hint = stringResource(R.string.areas_back_height_hint),
-                value = settings.backActivationHeightPercent,
-                onValueChangeFinished = { viewModel.setBackActivationHeight(it) },
-            )
-            SensitivitySlider(
-                label = stringResource(R.string.areas_back_sensitivity_label),
-                hint = stringResource(R.string.areas_back_sensitivity_hint),
-                baseDp = GestureZoneSettings.BASE_BACK_THICKNESS_DP,
-                value = settings.backEdgeSensitivity,
-                onValueChangeFinished = { viewModel.setBackEdgeSensitivity(it) },
-            )
-
-            HorizontalDivider(
-                modifier = Modifier.padding(horizontal = 16.dp, vertical = 4.dp),
-                color = MaterialTheme.colorScheme.outlineVariant.copy(alpha = 0.5f),
-            )
-
-            AreaSectionHeader(stringResource(R.string.areas_bottom_header))
-            Text(
-                text = stringResource(R.string.areas_bottom_shared_note),
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                modifier = Modifier.padding(horizontal = 16.dp),
-            )
-            PercentSlider(
-                label = stringResource(R.string.areas_bottom_width_label),
-                hint = stringResource(R.string.areas_bottom_width_hint),
-                value = settings.bottomActivationWidthPercent,
-                onValueChangeFinished = { viewModel.setBottomActivationWidth(it) },
-            )
-            SensitivitySlider(
-                label = stringResource(R.string.areas_bottom_sensitivity_label),
-                hint = stringResource(R.string.areas_bottom_sensitivity_hint),
-                baseDp = GestureZoneSettings.BASE_BOTTOM_THICKNESS_DP,
-                value = settings.bottomEdgeSensitivity,
-                onValueChangeFinished = { viewModel.setBottomEdgeSensitivity(it) },
-            )
-        }
-    }
-}
-
-@Composable
-private fun AreaSectionHeader(text: String) {
-    Text(
-        text = text,
-        style = MaterialTheme.typography.titleSmall.copy(fontWeight = FontWeight.SemiBold),
-        color = MaterialTheme.colorScheme.onSurface,
-        modifier = Modifier.padding(horizontal = 16.dp, vertical = 4.dp),
-    )
-}
-
-@Composable
-private fun PercentSlider(
-    label: String,
-    hint: String,
-    value: Int,
-    onValueChangeFinished: (Int) -> Unit,
-) {
-    // Local drag state snapped to the percentage step; persisted only when the thumb lifts.
-    var drag by remember(value) { mutableFloatStateOf(value.toFloat()) }
-    Column(modifier = Modifier.padding(horizontal = 16.dp, vertical = 4.dp)) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Text(
-                text = label,
-                style = MaterialTheme.typography.bodyMedium,
-                modifier = Modifier.weight(1f),
-            )
-            Text(
-                text = stringResource(R.string.areas_percent_value, drag.toInt()),
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-        Slider(
-            value = drag,
-            onValueChange = { drag = it },
-            onValueChangeFinished = {
-                // Round (don't truncate) to the nearest int before snapping to the 10% step,
-                // so a thumb released near 20% doesn't fall back to 10% via Float.toInt().
-                val snapped = GestureZoneSettings.clampPercent(Math.round(drag))
-                drag = snapped.toFloat()
-                onValueChangeFinished(snapped)
-            },
-            valueRange = GestureZoneSettings.PERCENT_MIN.toFloat()..GestureZoneSettings.PERCENT_MAX.toFloat(),
-            steps = (GestureZoneSettings.PERCENT_MAX - GestureZoneSettings.PERCENT_MIN) / GestureZoneSettings.PERCENT_STEP - 1,
-        )
-        Text(
-            text = hint,
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-    }
-}
-
-@Composable
-private fun SensitivitySlider(
-    label: String,
-    hint: String,
-    baseDp: Int,
-    value: Float,
-    onValueChangeFinished: (Float) -> Unit,
-) {
-    var drag by remember(value) { mutableFloatStateOf(value) }
-    Column(modifier = Modifier.padding(horizontal = 16.dp, vertical = 4.dp)) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Text(
-                text = label,
-                style = MaterialTheme.typography.bodyMedium,
-                modifier = Modifier.weight(1f),
-            )
-            val effectiveDp = (baseDp * drag).toInt().coerceAtLeast(1)
-            Text(
-                text = stringResource(R.string.areas_sensitivity_value, drag, effectiveDp),
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-        Slider(
-            value = drag,
-            onValueChange = { drag = it },
-            onValueChangeFinished = {
-                val snapped = GestureZoneSettings.clampSensitivity(drag)
-                drag = snapped
-                onValueChangeFinished(snapped)
-            },
-            valueRange = GestureZoneSettings.SENSITIVITY_MIN..GestureZoneSettings.SENSITIVITY_MAX,
-            steps = ((GestureZoneSettings.SENSITIVITY_MAX - GestureZoneSettings.SENSITIVITY_MIN) / GestureZoneSettings.SENSITIVITY_STEP).toInt() - 1,
-        )
-        Text(
-            text = hint,
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
     }
 }
